refactor: extract EmptyNodes.css.ts
